cf_sap_api_gateway/requirements.txt
google-cloud-secret-manager==2.12.6